% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{create_analysis}
\alias{create_analysis}
\title{Create a directory for standard AVALANCHE anlyses}
\usage{
create_analysis(
  path = ".",
  folder = "avalanche_analysis",
  name = "Author",
  title = "Untitled Analysis"
)
}
\arguments{
\item{path}{The directory path}

\item{folder}{The name of the new analysis project}

\item{title}{The title of the report}

\item{author}{The author's name}
}
\value{
invisibly, the path of the analysis directory
}
\description{
\code{create_analysis()} creates a project template that incudes \code{packages.R},
\code{analysis.R}, and \code{report.Rmd}.
}
