% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_resident_counts}
\alias{plot_resident_counts}
\title{Plot resident counts by sector}
\usage{
plot_resident_counts(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=get_resident_data]{get_resident_data()}}}
}
\value{
a ggplot
}
\description{
Plot resident counts by sector
}
\examples{

plot_resident_counts()

}
